requirements.txt
PySide6==6.5.0 
pyttsx3==2.90  
faster-whisper==0.10.0  
torch==2.0.1  
python-docx==0.8.11  
striprtf==0.0.22  
PyPDF2==3.0.1  
natasha==1.6.0  
transformers==4.38.2  
sentence-transformers==2.2.2  
llama-cpp-python==0.2.28 
pyaudio==0.2.14 